requirements.txt
Flask
joblib
scikit-learn
nltk
newspaper3k
pandas